Kill tab on request.

+++ background.js
@@ -12,14 +12,16 @@ chrome.runtime.onMessage.addListener(function(request, sender, sendResponse) {
   }
 });
 
-chrome.browserAction.onClicked.addListener(function(tab) {
-  queryInfo = {currentWindow: true, active : true};
-  chrome.tabs.query(queryInfo, function(result) {
-    var i;
-    for (i=0; i < result.length; i += 1) {
-      chrome.experimental.processes.getProcessIdForTab(result[i].id, function(processId) {
-        chrome.experimental.processes.terminate(processId);
-      });
-    }
-  });
+chrome.extension.onRequest.addListener(function(request, sender, sendResponse) {
+  if (request.action == "Avada Kedavra") {
+    queryInfo = {currentWindow: true, active : true};
+    chrome.tabs.query(queryInfo, function(result) {
+      var i;
+      for (i=0; i < result.length; i += 1) {
+        chrome.experimental.processes.getProcessIdForTab(result[i].id, function(processId) {
+          chrome.experimental.processes.terminate(processId);
+        });
+      }
+    });
+  }
 });
